requirements.txt
# Основные зависимости
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9

#pip install -r requirements.txt